pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-journalist"
version = "0.1.0"
description = "AI Journalist backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.10.3",
    "pydantic-settings==2.7.0",
    "langgraph==0.2.60",
    "langchain==0.3.13",
    "langchain-anthropic==0.3.3",
    "langchain-core>=0.3.28,<0.4.0",
    "langchain-community==0.3.13",
    "anthropic==0.42.0",
    "tavily-python==0.5.0",
    "playwright==1.49.1",
    "beautifulsoup4==4.12.3",
    "lxml==5.3.0",
    "httpx==0.28.1",
    "newsapi-python==0.2.7",
    "feedparser==6.0.11",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "psycopg2-binary==2.9.10",
    "aioboto3==13.2.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "google-api-python-client==2.154.0",
    "youtube-transcript-api==0.6.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "structlog==24.4.0",
    "tenacity==9.0.0",
    "aiofiles==24.1.0",
    "orjson==3.10.12",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-mock==3.14.0",
    "pytest-cov==6.0.0",
    "httpx==0.28.1",
    "anyio==4.7.0",
    "respx==0.21.1",
    "factory-boy==3.3.1",
    "aiosqlite==0.20.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["backend"]
omit = ["backend/migrations/*", "tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
